Finished code for removing a car from the stack
Finished code for removing a car from the stack. Allows users to remove a car from the stack, each element is removed from the driveway stack and moved to the street stack with both stacks being displayed each time until the number in the driveway stack matches with the  value the user inputted. Once this happens, that value is removed, both stacks are displayed, then the values from the street stack are moved back to the driveway stack with the stacks being displayed until the street stack is empty.